Xueqing version

+++ utils/tb_visualizer.py
@@ -22,9 +22,10 @@ class TBVisualizer:
         self._writer.close()
 
     def display_current_results(self, visuals, it, is_train, save_visuals=False):
+        # return
         for label, image_numpy in visuals.items():
             sum_name = '{}/{}'.format('Train' if is_train else 'Test', label)
-            self._writer.add_image(sum_name, image_numpy, it)
+            self._writer.add_image(sum_name, image_numpy.astype(np.uint8), it)
 
             if save_visuals:
                 util.save_image(image_numpy,
